+++ packages/muelu/adapters/epetra/MueLu_CreateEpetraPreconditioner.hpp
@@ -15,7 +15,7 @@
 #include <MueLu_MLParameterListInterpreter.hpp>
 #include <MueLu_ParameterListInterpreter.hpp>
 #include <MueLu_Utilities.hpp>
-#include <MueLu_HierarchyHelpers.hpp>
+#include <MueLu_HierarchyUtils.hpp>
 
 //! @file
 //! @brief Various adapters that will create a MueLu preconditioner that is an Epetra_Operator.
