+++ packages/muelu/adapters/stratimikos/Thyra_MueLuPreconditionerFactory_def.hpp
@@ -74,7 +74,7 @@
 
 #include <MueLu_Hierarchy.hpp>
 #include <MueLu_HierarchyManager.hpp>
-#include <MueLu_HierarchyHelpers.hpp>
+#include <MueLu_HierarchyUtils.hpp>
 #include <MueLu_ParameterListInterpreter.hpp>
 #include <MueLu_MLParameterListInterpreter.hpp>
 #include <MueLu_MasterList.hpp>
