+++ packages/muelu/src/Misc/MueLu_MergedBlockedMatrixFactory_def.hpp
@@ -59,7 +59,7 @@
 #include "MueLu_Level.hpp"
 #include "MueLu_Monitor.hpp"
 #include "MueLu_Utilities.hpp"
-#include "MueLu_HierarchyHelpers.hpp"
+//#include "MueLu_HierarchyHelpers.hpp"
 
 #include "MueLu_MergedBlockedMatrixFactory.hpp"
