+++ packages/muelu/src/Misc/MueLu_SchurComplementFactory_def.hpp
@@ -60,7 +60,7 @@
 #include "MueLu_Level.hpp"
 #include "MueLu_Monitor.hpp"
 #include "MueLu_Utilities.hpp"
-#include "MueLu_HierarchyHelpers.hpp"
+//#include "MueLu_HierarchyHelpers.hpp"
 
 #include "MueLu_SchurComplementFactory.hpp"
