+++ packages/muelu/src/Rebalancing/MueLu_RebalanceBlockInterpolationFactory_def.hpp
@@ -66,7 +66,7 @@
 #include "MueLu_RebalanceBlockInterpolationFactory_decl.hpp"
 
 #include "MueLu_FactoryManagerBase.hpp"
-#include "MueLu_HierarchyHelpers.hpp"
+#include "MueLu_HierarchyUtils.hpp"
 #include "MueLu_Level.hpp"
 #include "MueLu_Monitor.hpp"
 #include "MueLu_PerfUtils.hpp"
