+++ packages/muelu/src/Rebalancing/MueLu_RebalanceBlockRestrictionFactory_def.hpp
@@ -65,7 +65,7 @@
 
 #include "MueLu_RebalanceBlockRestrictionFactory_decl.hpp"
 
-#include "MueLu_HierarchyHelpers.hpp"
+#include "MueLu_HierarchyUtils.hpp"
 #include "MueLu_FactoryManagerBase.hpp"
 #include "MueLu_Level.hpp"
 #include "MueLu_Monitor.hpp"
