+++ packages/muelu/src/Smoothers/BlockedSmoothers/MueLu_BlockedDirectSolver_def.hpp
@@ -67,7 +67,7 @@
 #include "MueLu_Level.hpp"
 #include "MueLu_Utilities.hpp"
 #include "MueLu_Monitor.hpp"
-#include "MueLu_HierarchyHelpers.hpp"
+#include "MueLu_HierarchyUtils.hpp"
 #include "MueLu_SmootherBase.hpp"
 #include "MueLu_DirectSolver.hpp"
