+++ packages/muelu/src/Smoothers/BlockedSmoothers/MueLu_IndefBlockedDiagonalSmoother_def.hpp
@@ -68,7 +68,7 @@
 #include "MueLu_Level.hpp"
 #include "MueLu_Utilities.hpp"
 #include "MueLu_Monitor.hpp"
-#include "MueLu_HierarchyHelpers.hpp"
+#include "MueLu_HierarchyUtils.hpp"
 #include "MueLu_SmootherBase.hpp"
 #include "MueLu_SubBlockAFactory.hpp"
