+++ packages/muelu/src/Utils/ExplicitInstantiation/ETI_SC_LO_GO_NO_classes.cmake
@@ -31,6 +31,7 @@ APPEND_SET(MUELU_SC_LO_GO_NO_ETI_CLASSES MueLu::FactoryManager )
 APPEND_SET(MUELU_SC_LO_GO_NO_ETI_CLASSES MueLu::FilteredAFactory )
 APPEND_SET(MUELU_SC_LO_GO_NO_ETI_CLASSES MueLu::GenericRFactory )
 APPEND_SET(MUELU_SC_LO_GO_NO_ETI_CLASSES MueLu::Hierarchy )
+APPEND_SET(MUELU_SC_LO_GO_NO_ETI_CLASSES MueLu::HierarchyUtils )
 APPEND_SET(MUELU_SC_LO_GO_NO_ETI_CLASSES MueLu::Ifpack2Smoother-.?if.defined[HAVE_MUELU_TPETRA].&&.defined[HAVE_MUELU_IFPACK2] )
 APPEND_SET(MUELU_SC_LO_GO_NO_ETI_CLASSES MueLu::IndefBlockedDiagonalSmoother-.?ifdef.HAVE_MUELU_EXPERIMENTAL )
 APPEND_SET(MUELU_SC_LO_GO_NO_ETI_CLASSES MueLu::LineDetectionFactory )
@@ -78,6 +79,8 @@ APPEND_SET(MUELU_SC_LO_GO_NO_ETI_CLASSES MueLu::TentativePFactory_kokkos-.?if.de
 APPEND_SET(MUELU_SC_LO_GO_NO_ETI_CLASSES MueLu::ThresholdAFilterFactory )
 APPEND_SET(MUELU_SC_LO_GO_NO_ETI_CLASSES MueLu::ToggleCoordinatesTransferFactory )
 APPEND_SET(MUELU_SC_LO_GO_NO_ETI_CLASSES MueLu::TogglePFactory )
+APPEND_SET(MUELU_SC_LO_GO_NO_ETI_CLASSES MueLu::TopRAPFactory )
+APPEND_SET(MUELU_SC_LO_GO_NO_ETI_CLASSES MueLu::TopSmootherFactory )
 APPEND_SET(MUELU_SC_LO_GO_NO_ETI_CLASSES MueLu::TpetraOperator )
 APPEND_SET(MUELU_SC_LO_GO_NO_ETI_CLASSES MueLu::TransPFactory )
 APPEND_SET(MUELU_SC_LO_GO_NO_ETI_CLASSES MueLu::TrilinosSmoother )
