+++ packages/muelu/test/navierstokes/Navier2DBlocked_BraessSarazinSmoother.cpp
@@ -108,7 +108,7 @@
 #include "MueLu_Utilities.hpp"
 
 //TODO is it really needed?
-#include "MueLu_HierarchyHelpers.hpp"
+#include "MueLu_HierarchyUtils.hpp"
 
 #include <Epetra_LinearProblem.h>
 #include <AztecOO.h>
